Move IUserManager to common package
