+++ game-core/src/main/java/games/strategy/engine/lobby/common/IUserManager.java
@@ -1,9 +1,13 @@
-package games.strategy.engine.lobby.server;
+package games.strategy.engine.lobby.common;
 
 import games.strategy.engine.lobby.server.userDB.DBUser;
 import games.strategy.engine.message.IRemote;
 import games.strategy.engine.message.RemoteName;
 
+/**
+ * A service that provides user management operations. Currently, all operations may only target the calling user (i.e.
+ * {@code userName} must match the name of the calling user).
+ */
 public interface IUserManager extends IRemote {
   RemoteName USER_MANAGER =
       new RemoteName("games.strategy.engine.lobby.server.USER_MANAGER", IUserManager.class);
