+++ lobby/src/main/java/games/strategy/engine/lobby/server/UserManager.java
@@ -3,6 +3,7 @@ package games.strategy.engine.lobby.server;
 import org.mindrot.jbcrypt.BCrypt;
 
 import games.strategy.engine.config.lobby.LobbyPropertyReader;
+import games.strategy.engine.lobby.common.IUserManager;
 import games.strategy.engine.lobby.server.db.Database;
 import games.strategy.engine.lobby.server.db.HashedPassword;
 import games.strategy.engine.lobby.server.db.UserController;
